cdk/src/pipeline/extract/requirements.txt
numpy
requests
boto3
aws-xray-sdk
pillow